// app/lib/types.ts
export interface UserType {
  ID?: string;
  Username?: string;
  EmailAddress?: string;
  Password?: string;
  Active?: string;
  TimeNumber?: string;
  FirstName?: string;
  LastName?: string;
  Picture?: string;
  Gender?: string;
  Phone?: string;
  Country?: string;
  StageName?: string;
  Referrer?: string;
  isArtist?: string;
  isVerified?: string;
  isAdmin?: string;
  isManager?: string;
  Contract?: string;
  LatestNoteID?: string;
  RecordLabel?: string;
  IsPaused?: string;
  About?: string;
  IsArtist?: string;
  IsAdmin?:string;
  OTP?: string;
  Token?: string;

  // Optional fields (can be empty or null)
  OAuthProvider?: string;
  OAuthUID?: string;
  Locale?: string | null;
  Created?: string | null;
  Modified?: string | null;
  Link?: string | null;
  Facebook?: string | null;
  Twitter?: string | null;
  Instagram?: string | null;
  YouTube?: string | null;
  Suspension?: string | null;
  ReferralAmount?: string | null;
  Membership?: string | null;
  MembershipDeadline?: string | null;
  ActiveMembership?: string | null;
  IsServiceCharge?: string | null;
  ServiceChargeAmount?: string | null;
  IsMonthCharge?: string | null;
  MonthChargeAmount?: string | null;
  MonthChargePayTime?: string | null;
  MonthChargeExpireTime?: string | null;
  IsVerified?: string | null;
  Signature?: string | null;
  SignedContract?: string | null;
  Avatar?:string;
  WithdrawnPoints?: string | null;
  TotalPointsEarned?: string | null;
  PointRoyalty?: string | null;
  StripeCustomerID?: string | null;
  FCMToken?: string | null;
}
// ...
export type  navigationType  = {
  
  name: string;
  href: string;
  icon: any;
  hideOnMobile?: boolean;
  //action?: boolean|null;
  childOf?:string|null;
  parent?:string|null;
  };